pyproject.toml
[project]
name = "with4gent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "flask==3.1.2",
    "flask-cors==6.0.2",
    "line-bot-sdk==3.22.0",
    "openai==2.16.0",
    "pytest==9.0.2",
    "python-dotenv==1.2.1",
    "uvicorn>=0.40.0",
]
